rm modal view from about page
